rename: TaskDefinitionViewer.js to TaskDefinitionViewer.tsx
